refactor: db구조 변경에 따른 수정

+++ src/main/java/chess/dao/BoardDao.java
@@ -4,11 +4,11 @@ import chess.service.dto.BoardDto;
 import chess.service.dto.PieceWithSquareDto;
 
 public interface BoardDao {
-    void initBoard(int gameId);
+    void initBoard(Long gameId);
 
-    BoardDto getBoardByGameId(int id);
+    BoardDto getBoardByGameId(Long id);
 
-    void remove(int id);
+    void remove(Long id);
 
-    void update(PieceWithSquareDto piece, int gameId);
+    void update(PieceWithSquareDto piece, Long gameId);
 }
